refactor based on pr feedback
